+++ packages/core/styles/src/packages/atomic/text/nimbus-text.css.ts
@@ -69,6 +69,7 @@ const textColorProperties = {
 
   "success-surface": colorProperties["success-surface"],
   "success-interactive": colorProperties["success-textLow"],
+  "success-interactivePressed": colorProperties["success-interactivePressed"],
   "success-textLow": colorProperties["success-textLow"],
   "success-textHigh": colorProperties["success-textHigh"],